CategoriaDao + ServletCategoria + cambios realizados para que funcione categoria

+++ eCommerce/src/main/java/beans/LineaPedido.java
@@ -31,7 +31,7 @@ public class LineaPedido {
 		return producto;
 	}
 	
-	public double getPrecioLineaTotal() {
+	public int getPrecioLineaTotal() {
 		return cantidad * producto.getPrecio();
 	}
